+++ packages/core/babel-preset-dlight/src/pluginProvider.ts
@@ -14,7 +14,7 @@ import {
   defaultHTMLTags,
   devMode,
   dlightDefaultPackageName,
-  importList,
+  importMap,
 } from "./const"
 
 export class PluginProvider {
@@ -29,8 +29,6 @@ export class PluginProvider {
   private readonly excludes: string[]
   private readonly htmlTags: string[]
   private readonly attributeMap: Record<string, string[]>
-  private readonly importMap
-  private readonly minified: boolean
 
   constructor(
     babelApi: typeof babel,
@@ -39,8 +37,7 @@ export class PluginProvider {
     excludes: string[],
     enableDevTools: boolean,
     htmlTags: HTMLTags,
-    attributeMap: Record<string, string[]>,
-    minified: boolean
+    attributeMap: Record<string, string[]>
   ) {
     this.babelApi = babelApi
     this.t = types
@@ -57,13 +54,6 @@ export class PluginProvider {
             )
           : htmlTags
     this.attributeMap = attributeMap
-
-    this.importMap = Object.fromEntries(
-      importList.map((funcName, idx) =>
-        minified ? [funcName, funcName] : [funcName, `$${idx}$`]
-      )
-    )
-    this.minified = minified
   }
 
   // ---- DLight class Level
@@ -148,7 +138,7 @@ export class PluginProvider {
       // ---- Add nodes import to the head of file
       this.programNode!.body.unshift(
         this.t.importDeclaration(
-          Object.entries(this.importMap).map(([key, value]) =>
+          Object.entries(importMap).map(([key, value]) =>
             this.t.importSpecifier(
               this.t.identifier(value),
               this.t.identifier(key)
@@ -681,7 +671,7 @@ export class PluginProvider {
     const [body, classProperties, templateIdx] = generateView(viewParticles, {
       babelApi: this.babelApi,
       className: this.className!,
-      importMap: this.importMap,
+      importMap,
       subViewPropMap: Object.fromEntries(
         Object.entries(subViewPropSubDepMap).map(([key, props]) => [
           key,
@@ -690,7 +680,6 @@ export class PluginProvider {
       ),
       templateIdx: -1,
       attributeMap: this.attributeMap,
-      minified: this.minified,
     })
     viewNode.body = body
     this.classBodyNode?.body.push(...classProperties)
@@ -760,11 +749,10 @@ export class PluginProvider {
       {
         babelApi: this.babelApi,
         className: this.className!,
-        importMap: this.importMap,
+        importMap,
         subViewPropMap,
         templateIdx,
         attributeMap: this.attributeMap,
-        minified: this.minified,
       }
     )
     viewNode.body = body