+++ packages/core/transpiler/view-generator/src/types.ts
@@ -8,5 +8,4 @@ export interface ViewGeneratorConfig {
   subViewPropMap: SubViewPropMap
   templateIdx: number
   attributeMap: Record<string, string[]>
-  minified: boolean
 }